Project repls at last!
Just needed to add the nrepl dependency to the project so it was available to
start inside the new JVM, and then quote the require correctly.

I am able to (use 'overtone.live) and make a sound, which is great, but now it seems
the problem is with some of the connection machinery that is taking the output
from the other processes and printing it.